Moved stack into its own package
