+++ foth/stack/stack.go
@@ -1,4 +1,4 @@
-package main
+package stack
 
 import (
 	"fmt"
