+++ foth/stack/stack_test.go
@@ -1,4 +1,4 @@
-package main
+package stack
 
 import (
 	"testing"
